test: refactor utils
