--- test/util/sleep.js
@@ -1,9 +1,0 @@
-'use strict';
-
-async function sleep(ms) {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-}
-
-module.exports = sleep;
